Bluetooth Implementation Attempt
I started writing some code to implement Bluetooth. I couldn't test if it works because it requires creating a development build and installing it on an actual phone. Right now if you run it on Expo Go it causes an error if you go to settings because Expo Go cannot access Bluetooth.